自分の記事は、ブックマークできないよう修正

+++ app/controllers/bookmarks_controller.rb
@@ -4,6 +4,7 @@ class BookmarksController < ApplicationController
   before_action :authenticate_user!
   def create
     @article = Article.find(params[:article_id])
+    return if current_user?(@article.user)
     return unless @article.status == 'published'
 
     bookmark = @article.bookmarks.new(user_id: current_user.id)
@@ -13,6 +14,7 @@ class BookmarksController < ApplicationController
 
   def destroy
     @article = Article.find(params[:article_id])
+    return if current_user?(@article.user)
     bookmark = @article.bookmarks.find_by(user_id: current_user.id)
     bookmark.destroy if bookmark.present?
     redirect_to request.referer
