+++ app/controllers/likes_controller.rb
@@ -4,7 +4,9 @@ class LikesController < ApplicationController
   before_action :authenticate_user!
 
   def create
+    binding.pry
     @article = Article.find(params[:article_id])
+    return if current_user?(@article.user)
     return unless @article.status == 'published'
 
     like = @article.likes.new(user_id: current_user.id)
@@ -14,6 +16,7 @@ class LikesController < ApplicationController
 
   def destroy
     @article = Article.find(params[:article_id])
+    return if current_user?(@article.user)
     like = @article.likes.find_by(user_id: current_user.id)
     like.destroy if like.present?
     redirect_to request.referer
